requirements.txt
tensorflow==2.16.1
opencv-python==4.10.0.84
numpy==1.26.4
pandas==2.2.2
geopandas==0.14.4
folium==0.17.0
scikit-learn==1.5.1
matplotlib==3.9.0
streamlit==1.37.1
PyYAML==6.0.2
Pillow==10.4.0
